SWCDatasets

Documentation for SWCDatasets.

SWCDatasets.SourceData

SWCDatasets.SourceData

SWCDatasets.SourceData

SWCDatasets.SourceData

SWCDatasets.SourceData

SWCDatasets.SourceData

SWCDatasets.compress_save

SWCDatasets.compress_save!

SWCDatasets.dataset

SWCDatasets.dataset_table

SWCDatasets.datasets

SWCDatasets.dir_data

SWCDatasets.dir_raw

SWCDatasets.dir_to_be_converted

SWCDatasets.get_package_dataset_name

SWCDatasets.load_original

SWCDatasets.return_compressed

SWCDatasets.return_compressed

SWCDatasets.target_row

SWCDatasets.unzip_file

SWCDatasets.SourceData — Method

If zipfile not specified, it will be dir_data(package_name, dataset_name*".gz").

SWCDatasets.SourceData — Method

If rows, columns not specified, CSV.read(srcfile, DataFrame) will be applied to get the number of rows/columns.

SWCDatasets.SourceData — Method

If description not specified, it will be "".

SWCDatasets.SourceData — Method

SourceData(srcfile, packagename, datasetname, title, zipfile, rows, columns, description, timestamps)

srcfile is the path to the source file, the package_name will be the folder that the file resides, the dataset_name will be the name of the data without extension.

If timestamps not specified, it will be today().

SWCDatasets.SourceData — Method

If title not specified, it will be "Data [$dataset_name] of [$package_name]".

SWCDatasets.SourceData — Method

If package_name, dataset_name not specified, (package_name, dataset_name) = get_package_dataset_name(srcfile) is applied.

SWCDatasets.compress_save! — Method

compress_save!(SD::SourceData; move_source = true) compress the SD.srcfile, save the zipped one to SD.zipfile, and update the /home/runner/work/SWCDatasets.jl/SWCDatasets.jl/src/../data/doc/datasets.csv. By default, move_source = true that the source file will be moved to dir_raw().

SWCDatasets.compress_save — Method

compress_save(srcpath) is equivalent to compress_save!(SourceData(srcpath)) but returns SD::SourceData.

compress_save takes the same keyword arguments as compress_save!.

SWCDatasets.dataset — Method

dataset(package_name::AbstractString, dataset_name::AbstractString) returns a DataFrame object unzipped from the last row returned by target_row(package_name, dataset_name). This function mimics the dataset function in RDatasets.jl.

SWCDatasets.dataset_table — Method

The path to the index table for datasets in SWCDatasets. If SWCDatasets is added using pkg> dev SWCDatasets in other project/environment, dataset_table() returns "~/.julia/dev/SWCDatasets/src/../data/doc/datasets.csv".

SWCDatasets.datasets — Method

datasets() returns the table of this dataset, and define the global variable SWCDatasets.__datasets as this table.

Set ; force = true to force update SWCDatasets.__datasets with the dataset_table(); this keyword argument is intended to make some tests can work since in test dataset_table() is mutating. # todo: find a better way to deal with it.

SWCDatasets.dir_data — Method

Path to the directory for the data.

SWCDatasets.dir_raw — Method

Path to the directory for the backup of the raw data; only for package internal use.

SWCDatasets.dir_to_be_converted — Method

Path to the directory for the to-be-compressed data; only for package internal use.

SWCDatasets.get_package_dataset_name — Method

Given path to the source file, get_package_dataset_name(srcpath) derive package name and dataset name from the srcpath.

Example

srcpath = joinpath("Whatever", "RDatasets", "iris.csv")
SWCDatasets.get_package_dataset_name(srcpath)

# output

("RDatasets", "iris")

SWCDatasets.load_original — Method

load_original(path::AbstractString) opens path and return the read data.

SWCDatasets.return_compressed — Method

return_compressed(path::AbstractString) returned compressed data.

Example

compressed = return_compressed("data/data.csv")

SWCDatasets.return_compressed — Method

return_compressed(data::Vector{UInt8}) returned compressed data.

Example

data = load_original("data/data.csv")
compressed = return_compressed(data)

SWCDatasets.target_row — Method

target_row returns the latest information in datasets(). Given package_name, dataset_name, target_row(package_name, dataset_name), target_row returns the last row that matches row.PackageName == package_name && row.Dataset == dataset_name".

SWCDatasets.unzip_file — Method

The same as dataset, but also save the unzip file.